# perf(checker): add fetching metrics values in pipe