# Add fetching metrics values in pipe

Added fetching metrics to the redis pipeline

In addition, conducted benchmarks testing the performance of the solution on N number of unique metrics with 10 points each, the solution with pipeline greatly reduced the time to get values by metrics, but required more allocations

`old` - old solution without pipeline
`new` - new solution with pipeline

```
benchmark                                     old ns/op      new ns/op     delta
BenchmarkGetMetricValues/100-metrics-12       67507442       3536363       -94.76%
BenchmarkGetMetricValues/1000-metrics-12      512515118      26622719      -94.81%
BenchmarkGetMetricValues/5000-metrics-12      2600663168     94167600      -96.38%
BenchmarkGetMetricValues/10000-metrics-12     4881419523     179366564     -96.33%

benchmark                                     old allocs     new allocs     delta
BenchmarkGetMetricValues/100-metrics-12       5212           5126           -1.65%
BenchmarkGetMetricValues/1000-metrics-12      21133          51036          +141.50%
BenchmarkGetMetricValues/5000-metrics-12      158636         155034         -2.27%
BenchmarkGetMetricValues/10000-metrics-12     104532         147356         +40.97%

benchmark                                     old bytes     new bytes     delta
BenchmarkGetMetricValues/100-metrics-12       184949        182788        -1.17%
BenchmarkGetMetricValues/1000-metrics-12      833624        1845147       +121.34%
BenchmarkGetMetricValues/5000-metrics-12      5856144       5865810       +0.17%
BenchmarkGetMetricValues/10000-metrics-12     4961472       6296489       +26.91%
```